pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "industrial-ai-platform"
version = "0.1.0"
description = "Production-oriented industrial AI platform starter with FastAPI and MLOps scaffolding."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "numpy>=1.26,<3.0",
  "pillow>=10.0,<12.0",
  "pyspark>=3.5,<4.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27,<1.0",
  "pytest>=8.0,<9.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
